Add files via upload
PlayerUnknown's Battlegrounds (PUBG) is a popular online multiplayer battle royale game. In a battle royale game, players fight to be the last person or team standing. In order to win a PUBG game, it is important to have a combination of skill, strategy, and luck.

Now to predict the outcome of a game we need to train our model on a large dataset containing all various parameters of a game or information of a player like the guns he used, average headshot rate, his group rank etc. The datasets we are going to use 29 features. These 29 features along with the output of the game will be provided to train the model.